Fix interoperability with google SAML
